Creating database for saving last received result from api

+++ data/src/main/java/chistousov/ilya/data/api/dto/Wind.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class Wind(
     @SerializedName("deg") val deg: Int,
-    @SerializedName("speed") val speed: Int
+    @SerializedName("speed") val speed: Double
 )
